shop/requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.2.1
Sphinx==1.2.2
django-factory-boy==1.0.0
ipython==4.1.2